Don't look for a default id_asset.
Now that meta types are per-folder, to look up a default for id_asset
requires the item's meta types which requires the asset's meta types
which requires the id_asset which... causes an infinite loop.

Firefly uses an Item without an id_asset to represent "empty event"
and triggers this.